Move body generator for unit structs to its own module
Cut-paste
