+++ serde_derive/src/de.rs
@@ -19,6 +19,7 @@ mod enum_internally;
 mod enum_untagged;
 mod struct_;
 mod tuple;
+mod unit;
 
 pub fn expand_derive_deserialize(input: &mut syn::DeriveInput) -> syn::Result<TokenStream> {
     replace_receiver(input);
@@ -317,7 +318,7 @@ fn deserialize_body(cont: &Container, params: &Parameters) -> Fragment {
             Data::Struct(Style::Tuple, fields) | Data::Struct(Style::Newtype, fields) => {
                 tuple::deserialize_tuple(params, fields, &cont.attrs, TupleForm::Tuple)
             }
-            Data::Struct(Style::Unit, _) => deserialize_unit_struct(params, &cont.attrs),
+            Data::Struct(Style::Unit, _) => unit::deserialize_unit_struct(params, &cont.attrs),
         }
     } else {
         match &cont.data {
@@ -435,52 +436,6 @@ fn deserialize_try_from(type_try_from: &syn::Type) -> Fragment {
         _serde::#private::Result::and_then(
             <#type_try_from as _serde::Deserialize>::deserialize(__deserializer),
             |v| _serde::#private::TryFrom::try_from(v).map_err(_serde::de::Error::custom))
-    }
-}
-
-/// Generates `Deserialize::deserialize` body for a `struct Unit;`
-fn deserialize_unit_struct(params: &Parameters, cattrs: &attr::Container) -> Fragment {
-    let this_type = &params.this_type;
-    let this_value = &params.this_value;
-    let type_name = cattrs.name().deserialize_name();
-    let (de_impl_generics, de_ty_generics, ty_generics, where_clause) = params.generics();
-    let delife = params.borrowed.de_lifetime();
-
-    let expecting = format!("unit struct {}", params.type_name());
-    let expecting = cattrs.expecting().unwrap_or(&expecting);
-
-    quote_block! {
-        #[doc(hidden)]
-        struct __Visitor #de_impl_generics #where_clause {
-            marker: _serde::#private::PhantomData<#this_type #ty_generics>,
-            lifetime: _serde::#private::PhantomData<&#delife ()>,
-        }
-
-        #[automatically_derived]
-        impl #de_impl_generics _serde::de::Visitor<#delife> for __Visitor #de_ty_generics #where_clause {
-            type Value = #this_type #ty_generics;
-
-            fn expecting(&self, __formatter: &mut _serde::#private::Formatter) -> _serde::#private::fmt::Result {
-                _serde::#private::Formatter::write_str(__formatter, #expecting)
-            }
-
-            #[inline]
-            fn visit_unit<__E>(self) -> _serde::#private::Result<Self::Value, __E>
-            where
-                __E: _serde::de::Error,
-            {
-                _serde::#private::Ok(#this_value)
-            }
-        }
-
-        _serde::Deserializer::deserialize_unit_struct(
-            __deserializer,
-            #type_name,
-            __Visitor {
-                marker: _serde::#private::PhantomData::<#this_type #ty_generics>,
-                lifetime: _serde::#private::PhantomData,
-            },
-        )
     }
 }
